Handle removing LSB from code pointer in unwinder
The unwinder already has two mechanisms for converting what is found in
the LR to something useful.  There is `__builtin_extract_return_addr`
which takes the value from the LR (which can be found with
`__builtin_return_address`) and converts it into a relevant code
pointer, and there is `__builtin_frob_return_addr` which takes an
address and converts it into the value that will need to be in the LR or
stack slot for the epilogue to do its work.

These don't quite fit the need in Morello.  These act to adjust the
return address that the unwinder thinks about.  Hence
`__builtin_extract_return_addr` has been applied on the return value of
_Unwind_GetIP, and `__builtin_frob_return_addr` gets applied on the
value that we would give in _Unwind_GetIP in order to convert that value
back into what the epilogue needs.

In Morello the address found in the link register (or stack slot) is
sealed.  Hence if we were to use this same mechanism to adjust this
value, _Unwind_GetIP would now return invalid pointers, and
`__builtin_frob_return_addr` would not be able to make them valid again.
Hence we need a different mechanism.  The mechanism we choose is to
return the valid pointer including its LSB set in _Unwind_GetIP, and
then require any users of that pointer to remove the LSB if needed.  One
case where this may be needed is in the personality function of a
frame that may need the IP address (and not necessarily the full
pointer) in order to determine what action to take -- e.g. with C++
which landing pad if any is relevant for the current location.

In order to make this more convenient and to remove the architectural
knowledge required for its implementation, we introduce a new builtin
`__builtin_code_address_from_pointer`.  We use this in libgcc where it
is needed.  These places are:
 1. When finding the FDE associated with a location.
 2. When determining which CFA instructions to execute based on a
    location.

Future work using this new builtin in places like libbacktrace and the
personality functions will be done in future patches.

Small note on implementation:
 We expand the pointer into a Pmode value in the generic part of this
 builtin since the idea of the builtin is to return a code address
 instead of a pointer.   This is not relevant for Morello, but seems
 like it makes the operation of the function consistent across different
 architectures.

This patch also accounts for the Morello LSB in AArch64 unwind headers.
In aarch64-unwind.h we don't need to account for the Morello LSB, but
there is a function which looks like it might need it on the surface.
Here we add a comment to explain why it's not necessary in that case.

In freebsd-unwind.h our implementation of the fallback method for
finding the frame state (i.e. when there is no associated FDE with this
frame) tells whether this frame is a signal frame or not based on the
location of the PC.  This needs adjustment to account for the Morello
LSB and we adjust accordingly.
(This change has not been tested since we are not focussing on FreeBSD
for the moment).

In linux-unwind.h our fallback implementation works by dereferencing the
PC to tell if the code it points to matches the toolchains
implementation of __default_sa_restorer.  This method is no longer valid
for Morello since the PC is sealed and can not be dereferenced to read
in this way.  Hence we disable this fallback implementation for PureCap
Morello.
N.b. this does not disable unwinding past signal frames, only unwinding
past signal frames that do not have unwind information.  If the signal
frame has unwind information then it should include the 'S' character in
the augmentation string and we can find the frame state from the FDE
as usual.