Add unwinding tests to the testsuite
Our implementation of unwinding is as-yet only for C and hence does not
handle any LSDA/Personality function behaviour.

Our testcases ensure that _Unwind_ForcedUnwind and _Unwind_Backtrace
manage to unwind the entire stack, that the IP they find matches what we
expect for the part of the call-stack we create, and that the IP and CFA
they both find match each other.

We also check that the IP they find in one instance matches the IP found
by __builtin_return_addr in the function we use it, and that the CFA the
unwinder finds matches the frame address found by
__builtin_frame_address when that is available.

We test with both -fomit-frame-pointer and without.  Including the frame
pointer allows using `__builtin_frame_address` to validate the unwinder
behaviour, but removing the frame pointer should not invalidate the
unwinder information.  Hence we want to run the checks that should still
pass without a frame chain when using -fomit-frame-pointer.

N.b. the test using __builtin_frame_address is not foolproof, but that
we hope it doesn't exhibit its inherent flakyness.  We include a comment
in the testcase to that effect.