Comments around tricky DWARF expressions
There are three unwinder information directives which make use of DWARF
expressions.  These DWARF expressions describe an operation on a DWARF
stack machine.  The unwinder parses these operations in
`execute_cfa_program` and executes them in `execute_stack_op`.

There are three dwarf directives which make use of this abstract
machine: DW_CFA_expression, DW_CFA_val_expression, and
DW_CFA_def_cfa_expression.  These indicate that the following DWARF
machine operations calculate the location at which a register can be
found, the value that should be in a register, and the value that
represents the CFA of this frame respectively.

For these expressions to work with capabilities they must perform their
operations in such a way that the provenance is maintained.  For
example, binary operations performed on the top two items of the stack
must be performed by the execution routine in such a way that provenance
is taken from the item which the producer intended.  This means that
producer and consumer must agree.

As it happens the current implementation in libgcc takes provenance in
binary expressions from the earlier operand on the stack.  Similarly,
all expressions that GCC emits would want provenance to come from the
earlier operand on the stack -- see how the `and` expression is formed
for `build_cfa_aligned_loc`, and observe that the `REG_CFA_EXPRESSION`
note is only used for `reg+offset` and ends up just using instructions
which encode "register plus constant" directly (e.g. `DW_OP_fbreg`,
`DW_OP_breg<N>` operations).
The LLVM unwinder also takes provenance from the earlier operand, and
the idea seems to combine nicely with the fact that the evaluation of a
`DW_CFA_expression` and `DW_CFA_val_expression` start with an "initial
value" of the CFA placed on the DWARF stack (from which provenance seems
likely to be taken for `DW_CFA_expression`).

We make this decision more concrete and explicit by adjusting the types
of the first and second operand when evaluating binary operations in
`execute_stack_op`.

It should be mentioned that these expressions are rarely used.  GCC does
not emit `DW_CFA_val_expression` (even though it is implemented in the
backend no RTL emitted reaches this point) or
`DW_CFA_def_cfa_expression` (the codepath has an existing assertion that
we never hit this).  It is also impossible to emit `DW_CFA_expression`
for Morello at the moment (requirements on capabilities mean we would
not emit the RTL `(set sp (and sp <const_int>))` and we disable SVE for
Morello).
The LLVM unwinder does not implement `DW_CFA_val_expression` for
capabilities.

We tested this by temporarily modifying GCC to always emit the
`DW_CFA_expression` instructions that it could emit and ensuring that
the unwinder could correctly fetch the registers from the stack.  Since
the unwinder could fetch these registers the provenance was known to
have been preserved.