Remove LSB when finding landing pad correlating to a PC
Remove LSB when finding landing pad correlating to a PC

In the C personality function we need to find the landing pad associated
with a given PC.  To find the relevant landing pad we need to look for
the relevant PC location.   Since the PC we get from _Unwind_GetIP is a
valid sealed capability and removing that LSB loses the capability
validity we add code to maintain the clear difference between the IP
pointer and the IP as an address in the C personality function.

N.b. the implementation design of maintaining two variables is not
really necessary, we have no reason to need the IP from before.  However
it's still nice to maintain a clear distinction between the two ways
of representing IP.

N.b. this changes the cleanup-5.c test from FAIL to PASS because we now
correctly identify the LSDA for the instruction that we are in (rather
than the instruction that would come after the current branch).