Use executables PCC bounds with __EH_FRAME_BEGIN__
When passing this symbol to the register with the unwinder in
__register_frame_info we provide a capability that has this symbols
value with the metadata of PCC.  This means that the unwinder can use
this symbol as a base to take provenance from when accessing information
"outside" of the .eh_frame section.

This is only needed for exception unwinding code which uses personality
functions and landing pads.  Hence this is only needed for frames using
__attribute__((cleanup...)) with -fexceptions, or C++ with -fexceptions.

For the moment (i.e. this is not part of the final design) we find it
useful to keep the executable permissions from the PCC metadata.  This
is useful as it allows us to use the unwinder with pre-existing landing
pad information (which uses offsets to specify where the landing pad is
rather than requesting a pointer to a landing pad be provided by the
runtime and loading that).

In the future it would be best to remove these executable permissions
for security reasons.  The ABI for landing pads is needed to allow any
unwinder to iterate past frames in any objects (i.e. when they've not
been given any special pointer by the object).  When the specified ABI
is used there will be no benefit from having these executable
permissions and only extra security risk.