Temporary Morello hack for C only personality function
As a temporary step, we are using old landing-pad information which
specifies an offset from the call site record.  This offset does not
include the LSB needed for a pure capability ABI.  Here we add that LSB.

The final ABI records landing pad capabilities which are loaded by the
personality function.  These capabilities will come with the relevant
LSB.  Hence this change will not be needed for the final ABI.

Since this is a hack commit, also including a TODO comment about one of
the things we'll need to adjust when this gets cleaned up.