Allow some more block moves if we know it's safe according to the alignment
This fixes the second bug found in the compat.exp part
of the testsuite.
The compiler uses the block move mechanics for some
parts of parameter passing, like copying values from
one memory location to the stack when passing them by
value on the stack.
There was an edge case within this that would stop the
copy from happening completely, because `emit_block_move`
did not have any checking against `is_move_done` and it
would always assume that the required insn had been
emitted. It also had no other fall-back method of doing
the copy.
Here I have extended the block move logic to allow calls
to `emit_block_move_via_loop` if the alignment allows for
it, which means that we now do succeed in copying the
required data.
I have also added some checking to `emit_block_move`
and have added an assert so that we ICE if this ever
happens again.