Fortran: Fix KIND argument index for LEN_TRIM.
The mainline code to check whether an argument has to be included in
scalarization makes usage of a dummy argument object, which makes it possible
to rely on the dummy argument name only.  On the 11 branch, the dummy
argument object is not available and the code uses a mix of check for
argument name (for keyword arguments) and argument index (for non-keyword ones)
This makes backports non-straightforward in this area, as the argument indexes
depend on the intrinsics.

This change fixes a bogus backport for LEN_TRIM, which should have a different
KIND argument index from INDEX.
It also adds the missing MASKL, MASKR, SCAN and verify as backport from
master

	PR fortran/103789
	PR fortran/87711
	PR fortran/97896

gcc/fortran/ChangeLog:

	* trans-array.c (arg_evaluated_for_scalarization): Handle keyword and
	non-keyword arguments separatedly.  Adapt the expected argument index
	for KIND to each intrinsic in the non-keyword case.
	Add cases for MASKL, MASKR, SCAN and VERIFY.

gcc/testsuite/ChangeLog:

	* gfortran.dg/index_5.f90: Add check for non-keyword KIND argument.
	* gfortran.dg/len_trim.f90: Same.
	* gfortran.dg/maskl_1.f90: Same.
	* gfortran.dg/maskr_1.f90: Same.
	* gfortran.dg/scan_3.f90: Same.
	* gfortran.dg/verify_3.f90: Same.