Do not write known-bad capabilities out
This patch contains two parts.  The first is that when writing a
`capinit` or `chericap` relocation on a symbol plus addend where the
addend is known to be outside of the symbols range we do not use the
relevant directive but instead output an address with NULL metadata.
This change was made because the linker does not accept such `capinit`
directives -- it errors out.

We choose to warn yet still emit a bad capability rather than error out
since there is the chance this bad capability will never be used
depending on some other part of the code.  If a codebase would benefit
from erroring out then it can use -Werror.  We do not silently accept
such a feature since we judGE this would not usually be something
expected.

In order to account for this we will want to turn the error in GNU ld
into a warning.

For comparison, LLVM does not warn or error out on such invalid
capabilities.  The particular added in this patch emits a .capinit
directive for the base of the string and adds the invalid offset in
function code, but linking an object with LLD that has the CAPINIT
relocation with an offset past the bounds of the object it refers to is
still silently accepted.

This patch also includes changes to ensure alignment of anything that is
output as if it were a capability even if in reality the binary will not
end up with a capability.

This is needed since an `ldr` with a capability register requires a
16bit aligned address and will crash otherwise.  Usually the assembler
ensures alignment when it sees a `capinit` or `chericap` directive but
this obviously doesn't happen when that is not what we emit.
We care about this case since there is the possibility that a user will
generate an invalid capability and load it but be happy with an invalid
capability -- this could be because they do a test before using it or
are still happy with an invalid capability that has the address of
something.

In order to put this warning under the `-Wcheri-provenance` option we
made that `-Wcheri-provenance` option something accessible throughout
the compiler.

This change triggers a few new failures in the testsuite, these are all
because of the extra warnings acting as they should and hence triggering
new error messages that the testsuite did not expect.
The new failures are below, we fix up these by adding the appropriate
extra dg-warning directives.

gcc.dg/format/plus-1.c
gcc.dg/format/plus-1.c
gcc.dg/pr24683.c
gcc.dg/pr47727.c
gcc.dg/Warray-bounds-29.c
gcc.dg/Warray-bounds-31.c
gcc.dg/Warray-bounds-58.c

A note on implementation:
We put this in the AArch64 backend because we do not know what other
targets may or may not want to do in such situations.  Requesting an
out-of-bounds capability would be a valid choice.

Essentially the decision is being left for later when we have more
information about cases where this can come up.  All cases in the
testsuite so far do not appear to be cases where the valaue would be
expected to be dereferenced validly at runtime.