Update LSDA parsing of landing pads to handle the capability indirection.
This change parses the indirect offset method introduced for capabilities so
that the .eh_frame section no longer needs to be .rel.ro but can be read-only
instead.

Rather than define a relocation to add a landing pad capability in the
.eh_frame section itself, the new ABI defines an offset to be added in the
debug information (which can be relocated at link-time) and that offset points
to where a landing pad capability will be added at load time.

This change parses the indicator that this is the case and proceeds to read the
offset and the capability said offset points to.

N.b. we only support this method of indirection and not the older method of
having the capabilities directly in the dwarf unwind information.