morello: Implement new eh_frame landing pad layout
Morello uses a new layout for the call site information in the
gcc_exception_table that uses a "capability marker" value and
either a capability or an offset to a capability (details in
the ABI DWARF extension are TBD).

This patch implements only the second "indirect" schema that
uses an offset to a capability, as that is the one that is
intended to be stabilised and used going forwards.

Section switching to output the capability in .data.rel.ro is
handled in-line, rather than being queued for the end of the file
(which is unfortunate, but changing it would require a much more
significant overhaul of the debug output)

The table now gets output like this:

	.section	.gcc_except_table,"a",@progbits
.LLSDA3:
	.byte	0xff
	.byte	0xff
	.byte	0x1
	.uleb128 .LLSDACSE3-.LLSDACSB3
.LLSDACSB3:
	.uleb128 .LEHB0-.LFB3
	.uleb128 .LEHE0-.LEHB0
	.uleb128 0xd
.Llpoff0:
	.section	.data.rel.ro,"aw"
.Llpcap0:
	.capinit	doit+(.L10-(.LFB3))
	.xword	0
	.xword	0
	.section	.gcc_except_table
	.8byte	.Llpcap0-.Llpoff0
	.uleb128 0
	.uleb128 .LEHB1-.LFB3
	.uleb128 .LEHE1-.LEHB1
	.uleb128 0
	.uleb128 0

(the first entry has a landing pad, the second doesn't have a landing pad)