Mark XXSPLTIW/XXSPLTIDP as prefixed -- PR 104136
If you compile module_advect_em.F90 with -Ofast -mcpu=power10, one module
is large enough that we can't use a single conditional jump to span the
function.  Instead, we have to reverse the condition, and do a conditional
jump around an unconditional branch.  It turns out when xxspltiw and
xxspltdp instructions were generated, they were not marked as being
prefixed (i.e. length of 12 bytes instead of 4 bytes).  This meant the
calculations for the branch length were off, which in turn meant the
assembler raised an error because it couldn't do the conditional jump.

The fix is to set the maybe_prefixed attribute so that insns with the type
'vecperm' might be prefixed.  Then in the code that optionally puts a 'p'
in front of the insn skip doing so for the permutes (i.e. load constant
with splat instruction).

gcc/
2022-01-20   Michael Meissner  <meissner@linux.ibm.com>

	PR target/104136
	* config/rs6000/rs6000.cc (rs6000_final_prescan_insn): Do not
	prepend a 'p' for xxspltiw and xxspltidp.
	* config/rs6000/rs6000.md (maybe_prefixed): Indicate that vector
	permute instructions might be prefixed.