morello: Allow C++ frontend dummy objects
The C++ frontend sometimes needs to build short-lived "dummy objects" when it
needs an instance of object but doesn't have a concrete one available. See
cp/tree.c:build_dummy_object.

It forms these by building a pointer to the desired object type with
CONVERT_EXPR of void_node. As it stands, tree.h:capability_args_valid
would reject this as we're trying to build a capability from a
non-capability-typed tree node (void_node) with CONVERT_EXPR. This patch
allows specifically this construct as a special-case for the C++
frontend to permit it to use the same representation for these objects.

This should be safe provided these objects are short-lived and never
escape outside the C++ frontend, which I believe to be the case.