Use system default for long double if not specified on PowerPC.
If the user did not specify a default long double format, use the long double
default for the build compiler for the long double default.  This patch will
allow compilers built on a distribution that has changed the 128-bit floating
point format to use the default used on the system.

I did a normal normal bootstrap and make check regression on a little
endian power9 system and there were no regressions.

In addition, I built a compiler where I configured the default to use IEEE
128-bit floating point for long double.  I then used that compiler to
build a bootstrap with this patch applied and I did not set the floating
point format.  I verified that the compiler built with this patch defaults
long double to be IEEE 128-bit.

Can I apply this patch to the trunk for GCC 12?

gcc/
2022-01-26  Michael Meissner  <meissner@the-meissners.org>

	* config/rs6000/rs6000.cc (TARGET_IEEEQUAD_DEFAULT): If the
	compiler used to build the current compiler defaults to IEEE
	128-bit long double,  make that the default for this build.