cp: Some initial handling of capability conversions
The change in convert_to_integer_1 enables calling drop_capability on a
capability of NULLPTR_TYPE. We haven't seen this to date as this type is
only used by the C++ frontend.

The change in cp_build_binary_op ensures we compare the address value of
capability operands (as opposed to the capabilities themselves) and the
change in cp_build_unary_op ensures we use a non-capability type for the
increment amount.

gcc/ChangeLog:

	* convert.c (convert_to_integer_1): Also handle NULLPTR_TYPE.

gcc/cp/ChangeLog:

	* typeck.c (cp_build_binary_op): For comparisons, ensure we use a
	non-capability type as the common type for the arguments.
	(cp_build_unary_op): Ensure we use a non-capability type for the
	increment amount (since this is an offset).