Use correct names for __ibm128 if long double is IEEE 128-bit.
If you are on a PowerPC system where the default long double is IEEE
128-bit, GCC will use the wrong names for some of the conversion functions
for the __ibm128 type.

What is happening is when the defult long double is IEEE 128-bit, the
various convert, truncation, and extend functions did not specify a
default name for the conversion.  The machine indepentent portions of the
compiler would construct a call with an 'if' name (IFmode being the mode
for IBM 128-bit floating point).  This patch specifies to use the
tradiational 'tf' name for these conversion functions.

2022-01-28  Michael Meissner  <meissner@the-meissners.org>

gcc/
	PR target/104253
	* config/rs6000/rs6000.cc (TARGET_IEEEQUAD_DEFAULT): If the
	compiler used to build the current compiler defaults to IEEE
	128-bit long double,  make that the default for this build.

gcc/testsuite/
	PR target/104253
	* gcc.target/powerpc/pr104253.c: New test.