cp: Handle various capability conversions
This patch implements:
 - Casting from integers to capability pointers.
 - Implicit conversions from integer to intcap.
 - Handle pointer to intcap conversions.

Here we make a decision on where INTCAP_TYPE fits into the hierarcy set
out in cp-tree.h of CP_INTEGRAL_TYPE_P < ARITHMETIC_TYPE_P <
SCALAR_TYPE_P. With the initial C work, we made the decision that
INTCAP_TYPEs will not be INTEGRAL_TYPE_P. Since the types satisfying
CP_INTEGRAL_TYPE_P are a strict subset of INTEGRAL_TYPE_P, it follows
that INTCAP_TYPEs should not be CP_INTEGRAL_TYPE_P.

ARITHMETIC_TYPE_P then extends CP_INTEGRAL_TYPE_P with real and complex
types. We choose not to add INTCAP_TYPE here, since it's not possible to
perform arithmetic on INTCAP_TYPEs directly. Instead, we add them to
SCALAR_TYPE_P, along with the other preexisting pointer types and such.

As follow-on work, it will be necessary to audit places using
[CP_]INTEGRAL_TYPE_P and ARITHMETIC_TYPE_P to ensure that either that is
unreachable with INTCAP_TYPE or has the correct behaviour for
INTCAP_TYPE.

Notes on implementation: I considered using c_common_cap_from_noncap,
but the routine needs to use convert internally for an integer
conversion, and it made more sense to use the cp_convert function for
C++, and we also want to pass around the "complain" flags for C++, so at
this point defining a C++-specific routine seemed to make more sense.

gcc/cp/ChangeLog:

	* call.c (standard_conversion): Handle implicit conversions from
	integer to intcap.
	* cp-tree.h (SCALAR_TYPE_P): Return true for INTCAP_TYPEs.
	* cvt.c (cp_cap_from_noncap): New. Use it ...
	(cp_convert_to_pointer): ... here, and ...
	(ocp_convert): ... here (for converting to INTCAP_TYPE).
	* typeck.c (build_reinterpret_cast_1): Allow reinterpret_casting
	between suitably-sized integers and INTCAP_TYPE.