Use default long double from host compiler if not specified on PowerPC.
If the user did not specify a default long double format, set the long
double type to IEEE 128-bit if the host compiler uses IEEE 128-bit as the
long double default.

I used a Fedora rawhide system that was configured to use the default for
long double being IEEE 128-bit.  I built and bootstrapped a compiler on
this system with these patches applied, and I verified that the default
for the compiler used IEEE 128-bit long doubles.

Can I apply this patch to the trunk for GCC 12?  Ideally, this should be
back-ported to GCC 11 as well.

2022-02-02  Michael Meissner  <meissner@the-meissners.org>

gcc/
	* config.in (TARGET_POWERPC_IEEE_128BIT_LONG_DOUBLE): New
	configure option.
	* config/rs6000/rs6000.cc (TARGET_IEEEQUAD_DEFAULT): If the host
	compiler used IEEE 128-bit long doubles and the long double format
	was not specified with configuration switches, set the long double
	format to IEEE 128-bit.
	* configure.ac (TARGET_POWERPC_IEEE_128BIT_LONG_DOUBLE): Set as 1
	if the host compiler used IEEE 128-bit long doubles.
	* configure: Regenerate.