aarch64: Fix segfault in aarch64_mangle_type
For pure-capability Morello with the new varargs PCS, we just use void*
for __builtin_va_list and don't bother building the va_list_type node in
the backend.

This was causing a segfault via aarch64_mangle_type as it was calling
lang_hooks.types_compatible_p on a NULL va_list_type node. We don't want
special mangling for __builtin_va_list with the new PCS. It is simply
typedef'd to void * and should be mangled as such.  So this patch just
guards the special mangling on !TARGET_CAPABILITY_PURE, thereby also
fixing the segfault.

This patch allows us to compile some C++ code for purecap Morello.