Fix bugs around automatic capability alignment
In commit a77f6664310 we always aligned things that were being written
out as capabilities.  This triggered some failures in the testsuite,
which pointed out that the idea was mistaken.

The idea was based on the fact that loading an unaligned capability
triggers a fault hence thinking that we would only ever want to write
out aligned capabilities.  The choice to allow unaligned capabilities
being written out had already been made the other way, and that makes
more sense.  Allowing unaligned capabilities means that packed
structures with pointers in them are allowed (while we have a warning
on the unaligned capability).

This patch reverts that part of the commit, and also ensures we never use
`capinit` or `chericap` to emit capabilities that we want unaligned.

This adjustment is necessary since the assembler adds alignment when it sees
either of these directives.  Hence when trying to emit an unaligned capability
we need to emit it without these directives.

That change to emit unaligned capabilities with a different directive uncovered
an existing bug from the "precisely bounded capabilities" implementation (see
2167d4d7c41).  A few places mistakenly passed the alignment to the new hook in
bits, while the hook takes alignment in bytes, while others used the return
value from the new hook as if it were bits when it was in fact bytes.

We do not change the definition of the hook, since the units of the
corresponding hook are also bytes and we feel that matching these units still
seems like the least error-prone approach.

Here we simply use the helper function `alignment_pad_from_bits` that uses bits
for this alignment argument.

This patch does not come with a testsuite change, it is tested using the
existing testcases which emit unaligned capabilities and/or those which emit
aligned capabiilies going through those codepaths which have been modified.
(tests in the testsuite which have been seen to be affected by broken
implementations of this patch are:
 - gcc.c-torture/execute/20010915-1.c
 - gcc.c-torture/execute/920501-5.c
 - gcc.c-torture/execute/980223.c
 - gcc.c-torture/execute/builtins/strpbrk.c
 - gcc.dg/20000906-1.c
 - gcc.dg/torture/pr52530.c
 - gcc.dg/torture/pr53908.c
 - gcc.target/aarch64/aapcs64/morello-test_8.c
 - gcc.target/aarch64/morello/label-addressing-includes-lsb.c)

For ease of review, a description of the changes:
- assemble_noswitch_variable takes its alignment in bits hence we should pass
  the alignment argument of targetm.data_padding_size correspondingly adjusted.
  (can see this alignment is given in bits by the rest of the function).
- assemble_constant_contents mentions that it takes alignment in bits in the
  comment above the function.
- output_constant_def_contents passes its alignment to
  assemble_constant_contents hence its internal align variable should stay
  represented as bits.
- output_constant_pool_1 passes its `align_used` internal variable to
  `assemble_align` hence that should be in bits.
- place_block_symbol already did its calculations in bits, but we have a
  helper function which handles the case when there was a bitwise alignment and
  targetm.data_alignment did not change the alignment.  Hence we change the
  code to use that helper function.