Fix use of unmangled names in unwinding info
Commit dbc45b10e6 introduced assembly failures for purecap and fakecap
by emitting unmangled names in assembly. As it stands, the code builds a
SYMBOL_REF of current_function_name (), but this is the raw unmangled
name and is not suitable for use in assembly. This patch simply changes
the call to get_fnname_from_decl (current_function_decl) which gets the
mangled name from the decl's DECL_RTL.

Note that the code later calls switch_to_exception_section (name), and
indeed by inspecting other call sites of this function in the file, we
can see that they get the function name using the pattern above (with
get_fnname_from_decl).

There is a separate issue where we use the new exception table scheme
even for fakecap (we should just do it for purecap), but that should be
fixed separately.