Minor fix to exclude fake-capability from using the purecap exception table layout
During recent C++ work it was found that the "fake-capability" ABI
was picking up recent changes to the exception table landing pad
layout, that were intended only for Purecap. That is wrong, because
"fake-capability" is intended to always output standard A64 AAPCS
code, with the notion of capabilities existing only within the compiler.

This patch makes use of the `capabilities_in_hardware` target hook
to exclude this behaviour.