cp: Fix comparison of PTRMEMFUNCs with capabilities
When comparing the function pointers contained in two
pointer-to-member-functions, we must drop the capabilities to compare
the address values of the pointers. Note that we only need to adjust the
call to build2 where we're building the comparison directly: when going
through cp_build_binary_op we recursively ensure that the comparison is
done on the address values.

gcc/cp/ChangeLog:

	* typeck.c (cp_build_binary_op): When comparing PTRMEMFUNCs,
	ensure we compare the address values of the function pointer
	members.