value-prof: Fix invalid gimple created in gimple_ic
Prior to this patch, value-prof.c:gimple_ic could generate invalid
gimple by calling fold_drop_capability on the gimple_call_fn of a
gimple_call and using the result in a gimple_assign. The resulting tree
expression (in the case where a conversion was necessary) may be too
complicated to appear on the RHS of a gimple_assign.

This was causing an ICE due to invalid gimple when bootstrapping with
-mfake-capability, in particular when compiling tree-diagnostic-path.c.

This patch fixes the issue by introducing a new helper function
(gimple_drop_capability) in gimple-fold.h which can be used to drop the
capability on a gimple value operand, and adjusting the code in
value-prof.c to first load the gimple_call_fn into a pointer-typed
temporary. We then call the new helper function on the resulting
SSA_NAME and use the resulting address value in the comparison.

gcc/ChangeLog:

	* gimple-fold.c (gimple_drop_capability): New.
	* gimple-fold.h (gimple_drop_capability): New. Use it ...
	* value-prof.c (gimple_ic): ... here.